# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0-alpha.1
  description: Agentic Company OS API
servers:
  - url: /api
    description: Base API path
security:
  - operatorBearer: []
  - operatorSession: []
...
paths:
  /source-changes:
    get:
      operationId: listSourceChanges
      tags: [source]
      responses:
        "200":
          description: Source change history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceChange"
    post:
      operationId: prepareSourceChange
      tags: [source]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [id, agentId, sourcePath, request]
              properties:
                id: { type: string, format: uuid }
                agentId: { type: integer, minimum: 1 }
                sourcePath: { type: string, minLength: 1, maxLength: 2048 }
                request: { type: string, minLength: 1, maxLength: 4000 }
      responses:
        "200":
          description: Isolated source workspace and agent project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceChange"
  /source-changes/{id}:
    get:
      operationId: inspectSourceChange
      tags: [source]
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses:
        "200":
          description: Source change and redacted diff
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceChange"
  /source-changes/{id}/check:
    post:
      operationId: checkSourceChange
      tags: [source]
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expectedRevision, command]
              properties:
                expectedRevision: { type: integer, minimum: 1 }
                command:
                  oneOf:
                    - type: array
                      minItems: 1
                      maxItems: 64
                      items: { type: string }
                    - type: array
                      minItems: 1
                      maxItems: 4
                      items:
                        type: array
                        minItems: 1
                        maxItems: 64
                        items: { type: string }
      responses:
        "200":
          description: Recorded result; an unknown outcome must not be replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceChange"
        "409":
          description: State, revision, repository or permission conflict
  /source-changes/{id}/apply:
    post:
      operationId: applySourceChange
      tags: [source]
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expectedRevision]
              properties:
                expectedRevision: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Recorded result; an unknown outcome must not be replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceChange"
        "409":
          description: State, revision, repository or permission conflict
  /source-changes/{id}/rollback:
    post:
      operationId: rollbackSourceChange
      tags: [source]
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expectedRevision]
              properties:
                expectedRevision: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Recorded result; an unknown outcome must not be replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceChange"
        "409":
          description: State, revision, repository or permission conflict
  /skills/extensions:
    get:
      operationId: listPersonalCapabilities
      tags: [skills]
      summary: List persistent personal skills and utility presets
      responses:
        "200":
          description: Operator-owned installations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PersonalCapability"
    put:
      operationId: savePersonalCapability
      tags: [skills]
      summary: Atomically create or update a personal capability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [manifest, enabled, expectedRevision]
              properties:
                manifest:
                  $ref: "#/components/schemas/PersonalCapabilityManifest"
                enabled:
                  type: boolean
                expectedRevision:
                  type: integer
                  minimum: 0
      responses:
        "200":
          description: Saved capability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalCapability"
        "400":
          description: Invalid manifest
        "409":
          description: Reload the current revision before editing again
  /skills/extensions/{id}/export:
    get:
      operationId: exportPersonalCapability
      tags: [skills]
      summary: Export the manifest without installation state
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: ^user-[a-z0-9][a-z0-9-]*$
      responses:
        "200":
          description: Capability manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalCapabilityManifest"
        "404":
          description: Capability not found
  /skills/packs:
    get:
      operationId: getCapabilityPacks
      tags: [skills]
      summary: Read enabled capability packs
      responses:
        "200":
          description: Current selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapabilityPackSelection"
    put:
      operationId: updateCapabilityPacks
      tags: [skills]
      summary: Persist the selected packs with a revision check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enabledPacks, expectedRevision]
              properties:
                enabledPacks:
                  type: array
                  maxItems: 5
                  uniqueItems: true
                  items:
                    $ref: "#/components/schemas/CapabilityPackId"
                expectedRevision:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: Saved selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapabilityPackSelection"
        "400":
          description: Invalid selection
        "409":
          description: Revision conflict
  /skills:
    get:
      operationId: getCapabilityCatalog
      tags: [skills]
      summary: Read the built-in skills and local utility catalog
      parameters:
        - name: locale
          in: query
          schema:
            type: string
            enum: [tr, en, de, ru, zh-CN, zh-TW, ar]
            default: tr
      responses:
        "200":
          description: Versioned localized catalog; reading it does not start work or grant access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapabilityCatalog"
        "400":
          description: Invalid or repeated locale parameter
        "401":
          description: Operator authentication required
...
components:
...
  securitySchemes:
    operatorBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque operator token
    operatorSession:
      type: apiKey
      in: cookie
      name: agentic_os_session
  schemas:
    SourceChange:
      type: object
      required:
        [
          id,
          agentId,
          sourcePath,
          request,
          state,
          revision,
          baseCommit,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        agentId: { type: integer }
        sourcePath: { type: string }
        request: { type: string }
        state:
          type: string
          enum:
            [
              preparing,
              draft,
              checking,
              verified,
              applying,
              applied,
              rolling_back,
              rolled_back,
              failed,
              unknown,
            ]
        revision: { type: integer }
        baseCommit: { type: string }
        candidateCommit: { type: [string, "null"] }
        candidatePath: { type: [string, "null"] }
        appliedCommit: { type: [string, "null"] }
        taskId: { type: [integer, "null"] }
        error: { type: [string, "null"] }
        diff: { type: string }
        status: { type: string }
        workspace: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        check:
          type: [object, "null"]
          required: [command, exitCode, output, passed]
          properties:
            command: { type: array, items: { type: string } }
            commands:
              type: array
              items: { type: array, items: { type: string } }
            exitCode: { type: [integer, "null"] }
            output: { type: string }
            passed: { type: boolean }
    CapabilityPackId:
      type: string
      enum: [data, documents, web, code, planning]
    CapabilityPackSelection:
      type: object
      required: [id, enabledPacks, revision]
      properties:
        id:
          type: integer
          enum: [1]
        enabledPacks:
          type: array
          items:
            $ref: "#/components/schemas/CapabilityPackId"
        revision:
          type: integer
          minimum: 1
    PersonalCapabilityManifest:
      description: Versioned personal skill, bounded utility preset or operator-installed Node program. Maximum serialized size is 16000 characters. Programs declare terminal authority and execute through the exact approval-bound terminal path. No credential or arbitrary executable fields are accepted.
      oneOf:
        - type: object
          additionalProperties: false
          required: [schemaVersion, id, kind, title, description, instructions]
          properties:
            schemaVersion:
              type: integer
              enum: [1]
            id:
              type: string
              pattern: ^user-[a-z0-9][a-z0-9-]{0,59}$
            kind:
              type: string
              enum: [skill]
            title:
              type: string
              minLength: 1
              maxLength: 120
            description:
              type: string
              minLength: 1
              maxLength: 2000
            instructions:
              type: string
              minLength: 1
              maxLength: 8000
        - type: object
          additionalProperties: false
          required:
            [schemaVersion, id, kind, title, description, tool, defaults]
          properties:
            schemaVersion:
              type: integer
              enum: [1]
            id:
              type: string
              pattern: ^user-[a-z0-9][a-z0-9-]{0,59}$
            kind:
              type: string
              enum: [tool]
            title:
              type: string
              minLength: 1
              maxLength: 120
            description:
              type: string
              minLength: 1
              maxLength: 2000
            tool:
              type: string
              description: Exact built-in utility name from the capability catalog
            defaults:
              type: object
              additionalProperties: {}
        - type: object
          additionalProperties: false
          required:
            [schemaVersion, id, kind, title, description, code, permissions]
          properties:
            schemaVersion: { type: integer, enum: [1] }
            id: { type: string, pattern: "^user-[a-z0-9][a-z0-9-]{0,59}$" }
            kind: { type: string, enum: [program] }
            title: { type: string, minLength: 1, maxLength: 120 }
            description: { type: string, minLength: 1, maxLength: 2000 }
            code: { type: string, minLength: 1, maxLength: 8000 }
            permissions:
              type: array
              minItems: 1
              maxItems: 1
              items: { type: string, enum: [terminal] }
    PersonalCapability:
      type: object
      required: [id, manifest, enabled, revision, updatedAt]
      properties:
        id:
          type: string
        manifest:
          $ref: "#/components/schemas/PersonalCapabilityManifest"
        enabled:
          type: boolean
        revision:
          type: integer
          minimum: 1
        updatedAt:
          type: string
          format: date-time
...
    CapabilityLibraryCopy:
      type: object
      required:
        [
          title,
          intro,
          search,
          all,
          empty,
          skills,
          tools,
          inputs,
          steps,
          checks,
          deliverable,
          use,
          requirements,
          browser,
          files,
          boundary,
          toolBoundary,
          error,
          retry,
          loading,
          invalid,
          completed,
          skillMissing,
        ]
      properties:
        title:
          type: string
        intro:
          type: string
        search:
          type: string
        all:
          type: string
        empty:
          type: string
        skills:
          type: string
        tools:
          type: string
        inputs:
          type: string
        steps:
          type: string
        checks:
          type: string
        deliverable:
          type: string
        use:
          type: string
        requirements:
          type: string
        browser:
          type: string
        files:
          type: string
        boundary:
          type: string
        toolBoundary:
          type: string
        error:
          type: string
        retry:
          type: string
        loading:
          type: string
        invalid:
          type: string
        completed:
          type: string
        skillMissing:
          type: string
    BuiltinSkill:
      type: object
      required:
        [
          id,
          version,
          group,
          groupTitle,
          title,
          deliverable,
          inputs,
          steps,
          checks,
          tools,
          permissions,
        ]
      properties:
        id:
          type: string
        version:
          type: integer
        group:
          type: string
          enum: [research, engineering, data, content, operations]
        groupTitle:
          type: string
        title:
          type: string
        deliverable:
          type: string
        inputs:
          type: array
          items:
            type: string
        steps:
          type: array
          items:
            type: string
        checks:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
            enum: [canBrowse, canUseTerminal]
    CapabilityCatalog:
      type: object
      required: [locale, version, copy, groups, skills, tools]
      properties:
        locale:
          type: string
          enum: [tr, en, de, ru, zh-CN, zh-TW, ar]
        version:
          type: integer
        copy:
          $ref: "#/components/schemas/CapabilityLibraryCopy"
        groups:
          type: array
          items:
            type: object
            required: [id, title]
            properties:
              id:
                type: string
              title:
                type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/BuiltinSkill"
        tools:
          type: array
          items:
            type: object
            required: [name, title, description]
            properties:
              name:
                type: string
              title:
                type: string
              description:
                type: string
...