# lib/api-zod/src/generated/types/capabilityPackId.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Agentic Company OS API
 * OpenAPI spec version: 0.1.0-alpha.1
 */

export type CapabilityPackId = typeof CapabilityPackId[keyof typeof CapabilityPackId];


export const CapabilityPackId = {
  data: 'data',
  documents: 'documents',
  web: 'web',
  code: 'code',
  planning: 'planning',
} as const;