# lib/api-zod/src/generated/types/capabilityPackSelection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Agentic Company OS API
 * OpenAPI spec version: 0.1.0-alpha.1
 */
import type { CapabilityPackId } from './capabilityPackId';
import type { CapabilityPackSelectionId } from './capabilityPackSelectionId';

export interface CapabilityPackSelection {
  id: CapabilityPackSelectionId;
  enabledPacks: CapabilityPackId[];
  /** @minimum 1 */
  revision: number;
}